Move from version 1.5.1 to 1.5.2.
Issue solved - Pushdown did not work properly on all pages.

Added the ability to use up and down arrow to move back and forth, automatically focusing interact so it is easy for a use.  Up and down also move a user through sub menus.

Read on start as soon as the page is loaded.